# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "gowvp/owl HTTP API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/channels": {
      "get": {
        "summary": "通道列表",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "页码，从 1 开始"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "每页数量"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序字段，支持前缀 - 表示倒序"
            }
          },
          {
            "name": "did",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_online",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "has_recording",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "添加通道",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddChannel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/channels/{id}": {
      "put": {
        "summary": "修改通道",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditChannel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "删除通道",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/channels/{id}/recording-plan": {
      "get": {
        "summary": "查询通道录像计划",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelRecordingPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "绑定通道录像计划",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "required": [
                  "plan_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelRecordingPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "解绑通道录像计划",
        "tags": [
          "通道"
        ],
        "x-apifox-folder": "通道",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "unbound": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/recording-plans": {
      "get": {
        "summary": "录像计划列表",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordingPlan"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "新建录像计划",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingPlanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/recording-plans/{id}": {
      "get": {
        "summary": "录像计划详情",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新录像计划",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingPlanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "删除录像计划并解绑通道",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/recordings/play": {
      "get": {
        "summary": "查询录像播放地址",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "通道 ID",
              "example": "spjr0ms"
            }
          },
          {
            "name": "start_ms",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "开始时间戳（毫秒）",
              "example": 1789650000000
            }
          },
          {
            "name": "end_ms",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "结束时间戳（毫秒）",
              "example": 1789653600000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingPlay"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/recordings/timeline": {
      "get": {
        "summary": "查询录像时间轴",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "通道 ID",
              "example": "spjr0ms"
            }
          },
          {
            "name": "start_ms",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "开始时间戳（毫秒）",
              "example": 1789650000000
            }
          },
          {
            "name": "end_ms",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "结束时间戳（毫秒）",
              "example": 1789653600000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timeline"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/recordings/monthly": {
      "get": {
        "summary": "查询月份录像日期",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "通道 ID",
              "example": "spjr0ms"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 2026
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 9
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyStats"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/recordings": {
      "delete": {
        "summary": "删除时间范围内的录像",
        "tags": [
          "录像"
        ],
        "x-apifox-folder": "录像",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "通道 ID",
              "example": "spjr0ms"
            }
          },
          {
            "name": "start_ms",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "开始时间戳（毫秒）",
              "example": 1789650000000
            }
          },
          {
            "name": "end_ms",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "结束时间戳（毫秒）",
              "example": 1789653600000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRecordings"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddChannel": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "RTMP 或 RTSP"
          },
          "name": {
            "type": "string"
          },
          "device_id": {
            "type": "string"
          },
          "device_name": {
            "type": "string"
          },
          "app": {
            "type": "string"
          },
          "stream": {
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/StreamConfig"
          }
        }
      },
      "EditChannel": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ptztype": {
            "type": "integer",
            "format": "int32"
          },
          "is_online": {
            "type": "boolean"
          },
          "ext": {
            "$ref": "#/components/schemas/DeviceExt"
          },
          "app": {
            "type": "string"
          },
          "stream": {
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/StreamConfig"
          }
        }
      },
...
      "ChannelList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Channel"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "reason": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "details": {
                  "type": "object",
                  "properties": {}
                },
                "trace_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}